+++ resources/js/Components/Pages/Admin/AdminHeader.vue
@@ -5,8 +5,8 @@
         <h1 class="text-3xl font-semibold inline-block">
           <slot/>
         </h1>
-        <span class="text-white font-semibold badge badge-success absolute -top-2 -right-10 flex items-center justify-center min-w-[2rem] h-[2rem] rounded-full">{{ (badgePrimaryNumber ?? 0) > 99 ? '99+' : (badgePrimaryNumber ?? 0) }}</span>
-        <span class="font-semibold badge badge-warning absolute -top-2 -right-20 flex items-center justify-center min-w-[2rem] h-[2rem] rounded-full">{{ (badgeSecondaryNumber ?? 0) > 99 ? '99+' : (badgeSecondaryNumber ?? 0) }}</span>
+        <span v-if="displayBadges" class="text-white font-semibold badge badge-success absolute -top-2 -right-10 flex items-center justify-center min-w-[2rem] h-[2rem] rounded-full">{{ (badgePrimaryNumber ?? 0) > 99 ? '99+' : (badgePrimaryNumber ?? 0) }}</span>
+        <span v-if="displayBadges" class="font-semibold badge badge-warning absolute -top-2 -right-20 flex items-center justify-center min-w-[2rem] h-[2rem] rounded-full">{{ (badgeSecondaryNumber ?? 0) > 99 ? '99+' : (badgeSecondaryNumber ?? 0) }}</span>
 
       </div>
       <div>
@@ -25,9 +25,18 @@
 </template>
 
 <script setup>
-defineProps({
+import { reactive } from 'vue'
+
+const props = defineProps({
+  displayBadges: Boolean,
   badgePrimaryNumber: Number,
   badgeSecondaryNumber: Number,
-})
+});
+
+const defaultProps = {
+  displayBadges: false,
+};
+
+const finalProps = reactive({ ...defaultProps, ...props });
 
 </script>