+++ resources/js/Components/Pages/Admin/Channels/AdminChannelRow.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="table-row bg-white border-b dark:bg-gray-800 dark:border-gray-700">
     <div class="table-cell min-w-[8rem] px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap">
-      <span class="text-xl">{{ channel.name }}</span>
+      <button @click.prevent="openEditChannelModal" class="text-xl">{{ channel.name }}</button>
       <div v-if="channel?.isLive" class="ml-2 text-xs badge badge-secondary badge-outline">live</div>
     </div>
 
@@ -49,6 +49,17 @@
       </div>
     </div>
 
+    <dialog :id="'adminEditChannel_'+channel.id" class="modal">
+      <div class="modal-box">
+        <form method="dialog">
+          <button class="btn btn-sm btn-circle btn-ghost absolute right-2 top-2">✕</button>
+        </form>
+        <h3 class="font-bold text-lg mb-2">Edit Channel Name</h3>
+        <input v-model="channelName" type="text" class="input input-bordered w-full max-w-xs" />
+        <button @click.prevent="submit" class="btn btn-primary ml-2">Submit</button>
+      </div>
+    </dialog>
+
   </div>
 </template>
 
@@ -57,6 +68,7 @@ import { useAdminStore } from '@/Stores/AdminStore'
 import { useNotificationStore } from '@/Stores/NotificationStore'
 import AdminChannelDataItem from '@/Components/Pages/Admin/Channels/AdminChannelDataItem'
 import AdminChannelCurrentVideo from '@/Components/Pages/Admin/Channels/AdminChannelCurrentVideo.vue'
+import { computed, ref } from 'vue'
 
 const adminStore = useAdminStore()
 const notificationStore = useNotificationStore()
@@ -66,6 +78,18 @@ const props = defineProps({
 })
 
 const emit = defineEmits(['open-modal'])
+
+const channelName = ref(props.channel.name)
+
+const openEditChannelModal = () => {
+  document.getElementById('adminEditChannel_'+props.channel.id).showModal()
+}
+
+
+const submit = async () => {
+  await adminStore.updateChannel(props.channel.id, channelName.value)
+  document.getElementById('adminEditChannel_'+props.channel.id).close()
+}
 
 // Handle the open-modal event from grandchildren and re-emit to parent
 const handleOpenModal = (type) => {